% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N741pkg.R
\docType{package}
\name{N741pkg}
\alias{N741pkg}
\alias{N741pkg-package}
\title{N741pkg: A package for N741 course.}
\description{
N741pkg provides a few useful functions and a
template for reports and analyses in the N741
Big Data Analytics for Healthcare course
\url{https://melindahiggins2000.github.io/N741bigdata/}.
}
\section{N741pkg functions}{

The functions included are:

\itemize{
\item tbl.cat: creates a table of frequencies and percentages
for a categorical variable
\item tbl.continuous: creates a table of summary statistics
for a numeric variable
}
}

\section{Goals}{

The N741pkg package has the following goals:

\itemize{
\item to provide a Rmarkdown template.
\item goal 2.
\item goal 3.
}

To learn more about N741pkg, start with the vignettes:
`browseVignettes(package = "N741pkg")`
}

