% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.cat.R
\name{tbl.cat}
\alias{tbl.cat}
\title{Table of frequencies and percents for a categorical variable.}
\usage{
tbl.cat(gdf, var)
}
\arguments{
\item{gdf}{a grouped data.frame from dplyr::group_by().}

\item{var}{a vector in a data.frame df$var.}
}
\value{
\code{t} the frequency and percents table.
}
\description{
Create a table of summary statistics
for a categorical variable in a data frame.
}
\examples{
a <- runif(10,0,1)
b <- c("a","b","c","d","a","a","b","b","c","c")
c <- c("a",NA,"c","d","a","a","b","b","c",NA)
x <- data.frame(a,b,c)
gb <- dplyr::group_by(x,b)
gc <- dplyr::group_by(x,c)
tbl.cat(gb,gb$b)
tbl.cat(gc,gc$c)
}
