% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.continuous.R
\name{tbl.continuous}
\alias{tbl.continuous}
\title{Table of summary stats for a numeric variable.}
\usage{
tbl.continuous(df, var, l)
}
\arguments{
\item{df}{a data.frame.}

\item{var}{a vector in a data.frame df$var.}

\item{l}{a character vector for a label.}
}
\value{
\code{t} the summary statistics table.
}
\description{
Create a table of summary statistics
for a numeric variable in a data frame.
}
\examples{
a <- runif(10,0,1)
b <- c("a","b","c","d","a","a","b","b","c","c")
d <- c(1.1,2.2,3.3,NA,5.5,6.6,NA,8.8,9.9,NA)
x <- data.frame(a,b,d)
tbl.continuous(x,x$a,"label for var a")
tbl.continuous(x,x$d,"label for var d")
}
