% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N741pkg.R
\docType{package}
\name{N741pkg}
\alias{N741pkg}
\alias{N741pkg-package}
\title{N741pkg: A package for N741 course.}
\description{
N741pkg provides a few useful functions and a
template for reports and analyses in N741 course.
}
\details{
It has three main goals:

\itemize{
\item goal 1.
\item goal 2.
\item goal 3.
}

To learn more about N741pkg, start with the vignettes:
`browseVignettes(package = "N741pkg")`
}

